make save_as_df globally configurable
