+++ src/palantir/__init__.py
@@ -1,3 +1,4 @@
+from . import config
 from . import core
 from . import presults
 from . import io
@@ -7,6 +8,7 @@ from . import plot
 from .version import __version__
 
 __all__ = [
+    "config",
     "core",
     "presults",
     "io",
