+++ src/palantir/utils.py
@@ -13,101 +13,13 @@ import scanpy as sc
 
 from .core import run_palantir
 
+from .validation import _validate_obsm_key
+
 
 class CellNotFoundException(Exception):
     """Exception raised when no valid component is found for the provided cell type."""
 
     pass
-
-
-def _validate_obsm_key(ad, key, as_df=True):
-    """
-    Validates and retrieves the data associated with a specified key from the provided AnnData object.
-
-    Parameters
-    ----------
-    ad : sc.AnnData
-        The annotated data matrix from which the data is to be retrieved.
-    key : str
-        The key for accessing the data from the AnnData object's obsm.
-
-    Returns
-    -------
-    data : pd.DataFrame
-        A DataFrame containing the data associated with the specified key.
-    data_names : List[str]
-        A list of column names for the DataFrame.
-    as_df : bool, optional
-        If True, the data will be returned as pandas DataFrame with pseudotime as column names.
-        If False, the data will be returned as numpy array.
-        Default is True.
-
-    Raises
-    ------
-    KeyError
-        If the key or its corresponding columns are not found in the AnnData object.
-    """
-    if key not in ad.obsm:
-        raise KeyError(f"{key} not found in ad.obsm")
-    data = ad.obsm[key]
-    if not isinstance(data, pd.DataFrame):
-        if key + "_columns" not in ad.uns:
-            raise KeyError(
-                f"{key}_columns not found in ad.uns and ad.obsm[key] is not a DataFrame."
-            )
-        data_names = list(ad.uns[key + "_columns"])
-        if as_df:
-            data = pd.DataFrame(data, columns=data_names, index=ad.obs_names)
-    else:
-        data_names = list(data.columns)
-        if not as_df:
-            data = data.values
-    return data, data_names
-
-
-def _validate_varm_key(ad, key, as_df=True):
-    """
-    Validates and retrieves the data associated with a specified key from the provided AnnData object's varm attribute.
-
-    Parameters
-    ----------
-    ad : sc.AnnData
-        The annotated data matrix from which the data is to be retrieved.
-    key : str
-        The key for accessing the data from the AnnData object's varm.
-    as_df : bool, optional
-        If True, the trends will be returned as pandas DataFrame with pseudotime as column names.
-        If False, the trends will be returned as numpy array.
-        Default is True.
-
-    Returns
-    -------
-    data : Union[pd.DataFrame, np.ndarray]
-        A DataFrame or numpy array containing the data associated with the specified key.
-    data_names : List[str]
-        A list of column names for the DataFrame.
-
-    Raises
-    ------
-    KeyError
-        If the key or its corresponding columns are not found in the AnnData object.
-    """
-    if key not in ad.varm:
-        raise KeyError(f"{key} not found in ad.varm")
-    data = ad.varm[key]
-    if not isinstance(data, pd.DataFrame):
-        if key + "_pseudotime" not in ad.uns:
-            raise KeyError(
-                f"{key}_pseudotime not found in ad.uns and ad.varm[key] is not a DataFrame."
-            )
-        data_names = list(ad.uns[key + "_pseudotime"])
-        if as_df:
-            data = pd.DataFrame(data, columns=data_names, index=ad.var_names)
-    else:
-        data_names = list(data.columns)
-        if not as_df:
-            data = data.values
-    return data, data_names
 
 
 def run_pca(
